feat(observability): add shared OTLP package for both processes
- Moved the OTLP exporter and build metadata into a new @batuda/observability package so the mail-worker exports telemetry too, not just the server.
- The mail-worker now ships IMAP and session-fiber failures to Honeycomb and emits a throttled heartbeat for a liveness alert.
- The endpoint defaults unset so export stays off until Honeycomb is configured.
