+++ apps/server/package.json
@@ -36,6 +36,7 @@
 		"@batuda/domain": "workspace:*",
 		"@batuda/email": "workspace:*",
 		"@batuda/instructions": "workspace:*",
+		"@batuda/observability": "workspace:*",
 		"@batuda/research": "workspace:*",
 		"@batuda/ui": "workspace:*",
 		"@types/mailparser": "^3.4.6",
