+++ packages/observability/package.json
@@ -1,0 +1,30 @@
+{
+	"name": "@batuda/observability",
+	"version": "0.0.1",
+	"description": "Batuda — shared OTLP observability layer (traces, logs, metrics) + build metadata",
+	"type": "module",
+	"types": "./src/index.ts",
+	"main": "./src/index.ts",
+	"exports": {
+		".": {
+			"types": "./src/index.ts",
+			"import": "./src/index.ts"
+		}
+	},
+	"scripts": {
+		"build": "tsdown",
+		"check-types": "tsc --noEmit",
+		"test": "vitest run --passWithNoTests",
+		"test:watch": "vitest"
+	},
+	"dependencies": {
+		"effect": "4.0.0-beta.43"
+	},
+	"devDependencies": {
+		"@types/node": "^22.15.0",
+		"tsdown": "^0.21.7",
+		"typescript": "^5.9.3",
+		"vitest": "^4.1.4"
+	},
+	"private": true
+}
